Struct groupings::group::Group

pub struct Group { /* fields omitted */ }

Methods

impl Group

pub fn new(names: Vec<&str>) -> Group

Constructor taking a vector of strings and uses those as names for students.

pub fn set_name(&mut self, s: &str)

Name setter.

pub fn remove_sub(&mut self, ind: usize)

Remove a subgroup at an index and return all students in it to the group.

pub fn add_sub(&mut self, size: usize)

Add a subgroup of a specific size.

pub fn make_groups(&mut self)

Evaluate what students should be in what groups. (only use after creating groups with add_sub)

pub fn add_student(&mut self, name: &str)

Add a student

pub fn clear_subgroups(&mut self)

Clear subgroups vec.

pub fn fill_subs(&mut self)

Temporary. Fills subgroups.

pub fn finalise(&mut self) -> String

Finalise subgroups. Increment all weights.