Struct groupings::groupings::group::Group

pub struct Group { /* fields omitted */ }

Methods

impl Group

pub fn new(names: Vec<String>) -> Group

Constructor taking a vector of strings and uses those as names for students.

pub fn set_name(&mut self, s: String)

Name setter.

pub fn remove_sub(&mut self, ind: usize)

Remove a subgroup at an index and return all students in it to the group.

pub fn add_sub(&mut self, size: usize)

Add a subgroup of a specific size.

pub fn make_groups(&mut self)

Evaluate what students should be in what groups. (only use after creating groups with add_sub)

pub fn list_students(&self) -> &[Student]

For accessing the students in the class.