Struct groupings::student::Student

pub struct Student { /* fields omitted */ }

Methods

impl Student

pub fn id(&self) -> usize

Returns the ID as a usize

pub fn new(id: usize, name: &str, weight_ct: usize) -> Student

Student constructor. Takes an id, a name, and a number of weight slots to provide.

pub fn add_slot(&mut self)

Add an index to stor a student weight in.

pub fn get_weight(&self, p2: &Student) -> Option<u32>

Returns the weight of a grouping of p1 and p2, if they are the same. Otherwise, returns the higher weight or none.

pub fn set_weight(&mut self, p: &Student, weight: Option<u32>)

Setter for weight of a student.

pub fn inc_weight(&mut self, index: usize)

Increments weight by 1 at a given index.

pub fn set_pair_weight(p1: &mut Student, p2: &mut Student, weight: Option<u32>)

Sets a pair of students to a weight.

pub fn name(&self) -> String

Access the name field immutably.